requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
openpyxl>=3.1.0
plotly>=5.14.0
numpy>=1.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0
openai>=1.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
fpdf2>=2.7.0
